use correct type: size_t/int/long/pointer

+++ include/convert.h
@@ -6,6 +6,6 @@
 #define CONV_GB		1
 #define CONV_UTF8	2
 
-typedef int (* read_write_type)(int, void *, size_t);
+typedef ssize_t (* read_write_type)(int, void *, size_t);
 
 #endif
